u8 instead of long
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@47873 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ libpcsxcore/psxcommon.h
@@ -126,8 +126,8 @@ typedef struct {
 	boolean RCntFix;
 	boolean UseNet;
 	boolean VSyncWA;
-	long Cpu; // CPU_DYNAREC or CPU_INTERPRETER
-	long PsxType; // PSX_TYPE_NTSC or PSX_TYPE_PAL
+	u8 Cpu; // CPU_DYNAREC or CPU_INTERPRETER
+	u8 PsxType; // PSX_TYPE_NTSC or PSX_TYPE_PAL
 #ifdef _WIN32
 	char Lang[256];
 #endif
